+++ pkg/controller/federatedtypeconfig/controller.go
@@ -38,7 +38,7 @@ import (
 	"sigs.k8s.io/kubefed/pkg/controller/util"
 )
 
-const finalizer string = "core.kubefed.k8s.io/federated-type-config"
+const finalizer string = "core.kubefed.io/federated-type-config"
 
 // The FederatedTypeConfig controller configures sync and status
 // controllers in response to FederatedTypeConfig resources in the