+++ scripts/sync-up-helm-chart.sh
@@ -65,7 +65,7 @@ $SED -i 's/^metadata:/metadata:\n  annotations:\n    "helm.sh\/hook": crd-instal
 crd_diff=`(diff -U 4 ${TEMP_CRDS_YAML} ${CHART_FEDERATED_CRD_DIR}/crds.yaml; true;)`
 if [ -n "${crd_diff}" ]; then
   cp -f ${TEMP_CRDS_YAML} $CHART_FEDERATED_CRD_DIR/crds.yaml
-  $SED -i '1i{{ if (or (or (not .Values.global.scope) (eq .Values.global.scope "Cluster")) (not (.Capabilities.APIVersions.Has "core.kubefed.k8s.io\/v1beta1"))) }}' ${CHART_FEDERATED_CRD_DIR}/crds.yaml
+  $SED -i '1i{{ if (or (or (not .Values.global.scope) (eq .Values.global.scope "Cluster")) (not (.Capabilities.APIVersions.Has "core.kubefed.io\/v1beta1"))) }}' ${CHART_FEDERATED_CRD_DIR}/crds.yaml
   $SED -i '$a{{ end }}' ${CHART_FEDERATED_CRD_DIR}/crds.yaml
 fi
 
@@ -110,7 +110,7 @@ for filename in ./config/enabletypedirectives/*.yaml; do
   rm ${full_name}
 done
 $SED -i 's/^metadata:/metadata:\n  annotations:\n    "helm.sh\/hook": crd-install/'  ${CHART_FEDERATED_PROPAGATION_DIR}/crds.yaml
-$SED -i '1i{{ if (or (or (not .Values.global.scope) (eq .Values.global.scope "Cluster")) (not (.Capabilities.APIVersions.Has "types.kubefed.k8s.io\/v1beta1"))) }}' ${CHART_FEDERATED_PROPAGATION_DIR}/crds.yaml
+$SED -i '1i{{ if (or (or (not .Values.global.scope) (eq .Values.global.scope "Cluster")) (not (.Capabilities.APIVersions.Has "types.kubefed.io\/v1beta1"))) }}' ${CHART_FEDERATED_PROPAGATION_DIR}/crds.yaml
 $SED -i '$a{{ end }}' ${CHART_FEDERATED_PROPAGATION_DIR}/crds.yaml
 
 # Clean kube-apiserver daemons and temporary files